requirements.txt
python==3.10
pytorch==2.0.0
matplotlib==3.8.0